Add files via upload
This is a chrome extension that i have spent over weeks creating without a single godamn line of Jquery. If you dont know what Jquery is  its basically and old library of javascript. But i used vanilla javascript so it was way harder. So back to the point the extension basically modifies the official amazon page to let you fake purchase anything and it also modifies some stuff to make it look like you are in a account. In the extension it modifies the text to say "Hello, Pedro". My name is not pedro I was just putting a random name. So if you want to modifie the name just go to content.js in thee extension file and look for this line of code : $("nav-link-accountList-nav-line-1").textContent="Hello, Pedro" and change the Hello, Pedro to "Hello, (Whatever your name is)" and then just save the file and then unload it in your browser but obviously you first got to unzip it. One of the most important things about this extension is that you have to wait about 2 to 3 seconds after the page loads before clicking the buy button becuase the extension has to load. So if you didnt understand just wait until the name at the top right corner changes and then you can click the buy noe button and everything will work fine. Another idea that someone could create that i have been unable to create is making a whit div appear while the extension runs so it is not visible when the text changes because then it could seem suspicious but if you want to create that pls make a fork. Thank you for reading this if you did and if you didnt its fine but i hope you enjoy your trolling orr scambaiting and Merry Trolling for my trollers.